pyproject.toml
[project]
name = "fastapi-arch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "aiomysql>=0.2.0",
    "fastapi[standard]>=0.116.1",
    "greenlet>=3.2.4",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "redis>=6.4.0",
    "sqlmodel>=0.0.24",
]
